Module mynewt::encoding::coap_context

COAP encoder state used by CoAP encoding macros

Structs

CoapContext
CoAP encoder state. Buffers the next key and value to be encoded.

Enums

CoapError
Error codes for COAP encoding failure

Constants

COAP_KEY_SIZE
Size of the static key buffer

COAP_VALUE_SIZE
Size of the static value buffer

_ARRAY
Array element of CoAP document

_MAP
Map element of CoAP document

_ROOT
Root of CoAP document

Statics

COAP_CONTEXT
Global instance that contains the current state of the CoAP encoder. Only 1 encoding task is supported at a time.

cbor_encoder0
Global CBOR root map for encoding CBOR documents

cbor_encoder1

Traits

ToBytesOptionalNull
Convert the type to array of bytes that may or may not end with null