Struct mynewt::encoding::coap_context::CoapContext

pub struct CoapContext {
    key_buffer: [u8; 32],
    value_buffer: [u8; 32],
}

CoAP encoder state. Buffers the next key and value to be encoded.

Fields
key_buffer: [u8; 32]

Static buffer for the key to be encoded. Will be passed to Mynewt COAP encoder API. Always null-terminated.

value_buffer: [u8; 32]

Static buffer for the string value to be encoded. Will be passed to Mynewt COAP encoder API. Always null-terminated.

Methods

impl CoapContext

pub fn json_set_geolocation(
&mut self,
_key: &Strn,
_lat_key: &Strn,
_long_key: &Strn,
_geo: SensorValueType
)

pub fn json_set_text_string(&mut self, key: &Strn, value: &Strn)

Encode a text value into the current JSON document with the specified key

fn key_strn_to_cstr(&mut self, key: &Strn) -> *const u8

Given a Strn key key, return a *char pointer that is null-terminated. Used for encoding COAP keys. If key is null-terminated, return it as a pointer. Else copy key to the static key buffer, append null and return the static key buffer as a pointer.

fn value_strn_to_cstr(&mut self, value: &Strn) -> *const u8

Given a Strn value value, return a *char pointer that is null-terminated. Used for encoding COAP values. If value is null-terminated, return it as a pointer. Else copy value to the static value buffer, append null and return the static value buffer as a pointer.

pub fn key_to_cstr(&mut self, s: &[u8]) -> *const u8

Given a key s, return a *char pointer that is null-terminated. Used for encoding COAP keys. If s is null-terminated, return it as a pointer. Else copy s to the static key buffer, append null and return the static key buffer as a pointer.

pub fn value_to_cstr(&mut self, s: &[u8]) -> *const u8

Given a value s, return a *char pointer that is null-terminated. Used for encoding COAP values. If s is null-terminated, return it as a pointer. Else copy s to the static value buffer, append null and return the static value buffer as a pointer.

pub fn cstr_len(&self, s: &[u8]) -> usize

Compute the byte length of the string in s. If s is null-terminated, return length of s - 1. Else return length of s.

pub fn global_encoder(&self) -> *mut CborEncoder

Return the global CBOR encoder

pub fn new_encoder(&self, key: &str, suffix: &str) -> *mut CborEncoder

Create a new CBOR encoder for the current map or array, e.g. key=root, suffix=_map

pub fn encoder(&self, key: &str, suffix: &str) -> *mut CborEncoder

Return the CBOR encoder for the current map or array, e.g. key=root, suffix=_map

pub fn check_result(&self, res: u32)

Fail the encoding with an error if res is non-zero.

pub fn fail(&mut self, err: CoapError)

Fail the encoding with an error

pub fn to_void_ptr(&mut self) -> *mut c_void

Cast itself as a *mut c_void