Trait mynewt::encoding::coap_context::ToBytesOptionalNull

pub trait ToBytesOptionalNull {
    fn to_bytes_optional_nul(&self) -> &[u8];
}

Convert the type to array of bytes that may or may not end with null

Required methods

fn to_bytes_optional_nul(&self) -> &[u8]

Convert the type to array of bytes that may or may not end with null

Loading content...