Module mynewt::encoding::tinycbor

Contains Rust bindings for Mynewt TinyCBOR Encoding API encoding/tinycbor

Structs

CborEncoder

CborParser

CborValue

File

File_methods

cbor_decoder_reader

cbor_encoder_writer

cbor_iovec

Constants

CborError_CborErrorAdvancePastEOF

CborError_CborErrorDataTooLarge

CborError_CborErrorDuplicateObjectKeys

CborError_CborErrorGarbageAtEnd

CborError_CborErrorIO

CborError_CborErrorIllegalNumber

CborError_CborErrorIllegalSimpleType

CborError_CborErrorIllegalType

CborError_CborErrorInappropriateTagForType

CborError_CborErrorInternalError

CborError_CborErrorInvalidUtf8TextString

CborError_CborErrorJsonNotImplemented

CborError_CborErrorJsonObjectKeyIsAggregate

CborError_CborErrorJsonObjectKeyNotString

CborError_CborErrorNestingTooDeep

CborError_CborErrorOutOfMemory

CborError_CborErrorTooFewItems

CborError_CborErrorTooManyItems

CborError_CborErrorUnexpectedBreak

CborError_CborErrorUnexpectedEOF

CborError_CborErrorUnknownLength

CborError_CborErrorUnknownSimpleType

CborError_CborErrorUnknownTag

CborError_CborErrorUnknownType

CborError_CborErrorUnsupportedType

CborError_CborNoError

CborError_CborUnknownError

CborIndefiniteLength

CborKnownTags_CborBase64urlTag

CborKnownTags_CborBase64Tag

CborKnownTags_CborBigfloatTag

CborKnownTags_CborDateTimeStringTag

CborKnownTags_CborDecimalTag

CborKnownTags_CborExpectedBase16Tag

CborKnownTags_CborExpectedBase64urlTag

CborKnownTags_CborExpectedBase64Tag

CborKnownTags_CborMimeMessageTag

CborKnownTags_CborNegativeBignumTag

CborKnownTags_CborPositiveBignumTag

CborKnownTags_CborRegularExpressionTag

CborKnownTags_CborSignatureTag

CborKnownTags_CborUnixTime_tTag

CborKnownTags_CborUriTag

CborMajorTypes_ArrayType

CborMajorTypes_ByteStringType

CborMajorTypes_MapType

CborMajorTypes_NegativeIntegerType

CborMajorTypes_SimpleTypesType

CborMajorTypes_TagType

CborMajorTypes_TextStringType

CborMajorTypes_UnsignedIntegerType

CborParserIteratorFlags_CborIteratorFlag_ContainerIsMap

CborParserIteratorFlags_CborIteratorFlag_IntegerValueTooLarge

CborParserIteratorFlags_CborIteratorFlag_NegativeInteger

CborParserIteratorFlags_CborIteratorFlag_UnknownLength

CborSimpleTypes_Break

CborSimpleTypes_DoublePrecisionFloat

CborSimpleTypes_FalseValue

CborSimpleTypes_HalfPrecisionFloat

CborSimpleTypes_NullValue

CborSimpleTypes_SimpleTypeInNextByte

CborSimpleTypes_SinglePrecisionFloat

CborSimpleTypes_TrueValue

CborSimpleTypes_UndefinedValue

CborType_CborArrayType

CborType_CborBooleanType

CborType_CborByteStringType

CborType_CborDoubleType

CborType_CborFloatType

CborType_CborHalfFloatType

CborType_CborIntegerType

CborType_CborInvalidType

CborType_CborMapType

CborType_CborNullType

CborType_CborSimpleType

CborType_CborTagType

CborType_CborTextStringType

CborType_CborUndefinedType

Functions

cbor_encode_byte_iovec⚠
Appends the byte string passed as \a iov and \a iov_len to the CBOR stream provided by \a encoder. CBOR byte strings are arbitrary raw data.

cbor_encode_byte_string⚠
Appends the text string \a string of length \a length to the CBOR stream provided by \a encoder. CBOR requires that \a string be valid UTF-8, but TinyCBOR makes no verification of correctness.

cbor_encode_floating_point⚠
Appends the floating-point value of type \a fpType and pointed to by \a value to the CBOR stream provided by \a encoder. The value of \a fpType must be one of CborHalfFloatType, CborFloatType or CborDoubleType, otherwise the behavior of this function is undefined.

cbor_encode_int⚠
Appends the signed 64-bit integer \a value to the CBOR stream provided by \a encoder.

cbor_encode_negative_int⚠
Appends the negative 64-bit integer whose absolute value is \a absolute_value to the CBOR stream provided by \a encoder.

cbor_encode_simple_value⚠
Appends the CBOR Simple Type of value \a value to the CBOR stream provided by \a encoder.

cbor_encode_tag⚠
Appends the CBOR tag \a tag to the CBOR stream provided by \a encoder.

cbor_encode_text_string⚠
Appends the byte string \a string of length \a length to the CBOR stream provided by \a encoder. CBOR byte strings are arbitrary raw data.

cbor_encode_uint⚠
Appends the unsigned 64-bit integer \a value to the CBOR stream provided by \a encoder.

cbor_encoder_close_container⚠
Closes the CBOR container (array, map or indefinite-length string) provided by \a containerEncoder and updates the CBOR stream provided by \a encoder. Both parameters must be the same as were passed to cbor_encoder_create_array() or cbor_encoder_create_map() or cbor_encoder_create_indef_byte_string().

cbor_encoder_close_container_checked⚠

cbor_encoder_create_array⚠
Creates a CBOR array in the CBOR stream provided by \a encoder and initializes \a arrayEncoder so that items can be added to the array using the CborEncoder functions. The array must be terminated by calling either cbor_encoder_close_container() or cbor_encoder_close_container_checked() with the same \a encoder and \a arrayEncoder parameters.

cbor_encoder_create_indef_byte_string⚠
Creates a indefinite-length byte string in the CBOR stream provided by \a encoder and initializes \a stringEncoder so that chunks of original string can be added using the CborEncoder functions. The string must be terminated by calling cbor_encoder_close_container() with the same \a encoder and \a stringEncoder parameters.

cbor_encoder_create_indef_text_string⚠
Creates a indefinite-length text string in the CBOR stream provided by \a encoder and initializes \a stringEncoder so that chunks of original string can be added using the CborEncoder functions. The string must be terminated by calling cbor_encoder_close_container() with the same \a encoder and \a stringEncoder parameters.

cbor_encoder_create_map⚠
Creates a CBOR map in the CBOR stream provided by \a encoder and initializes \a mapEncoder so that items can be added to the map using the CborEncoder functions. The map must be terminated by calling either cbor_encoder_close_container() or cbor_encoder_close_container_checked() with the same \a encoder and \a mapEncoder parameters.

cbor_encoder_init⚠
Initializes a CborEncoder structure \a encoder by pointing it to buffer \a buffer of size \a size. The \a flags field is currently unused and must be zero.

cbor_error_string⚠

cbor_parser_init⚠

cbor_value_advance⚠

cbor_value_advance_fixed⚠

cbor_value_calculate_string_length⚠

cbor_value_enter_container⚠

cbor_value_get_half_float⚠

cbor_value_get_int64_checked⚠

cbor_value_get_int_checked⚠

cbor_value_leave_container⚠

cbor_value_map_find_value⚠

cbor_value_skip_tag⚠

cbor_value_text_string_equals⚠

cbor_value_to_pretty_advance⚠

Type Definitions

CborError

CborKnownTags

CborMajorTypes

CborParserIteratorFlags

CborSimpleTypes

CborTag

CborType

FILE

__int64_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uintptr_t

cbor_encoder_write

cbor_memcmp

cbor_memcpy

cbor_reader_get8

cbor_reader_get16

cbor_reader_get32

cbor_reader_get64