Function mynewt::hw::hal::hal_i2c_master_probe

pub unsafe extern "C" fn hal_i2c_master_probe(
    i2c_num: u8, 
    address: u8, 
    timeout: u32
) -> c_int

Probes the i2c bus for a device with this address. THIS API issues a start condition, probes the address using a read command and issues a stop condition.

i2c_num: The number of the I2C to probe

address: The address to probe for

timeout: How long to wait for transaction to complete in ticks

Return: 0 on success, non-zero error code on failure