Type Definition mynewt::hw::sensor::sensor_get_config_func_t

type sensor_get_config_func_t = Option<unsafe extern "C" fn(arg1: *mut sensor, arg2: sensor_type_t, arg3: *mut sensor_cfg) -> c_int>;

Get the configuration of the sensor for the sensor type. This includes the value type of the sensor.

sensor: Ptr to the sensor

type: The type of sensor value to get configuration for

cfg: A pointer to the sensor value to place the returned result into.

Return: 0 on success, non-zero error code on failure.