Type Definition mynewt::hw::sensor::sensor_set_notification_t

type sensor_set_notification_t = Option<unsafe extern "C" fn(arg1: *mut sensor, arg2: sensor_event_type_t) -> c_int>;

Set the notification expectation for a targeted set of events for the specific sensor. After this function returns successfully, the implementer shall post corresponding event notifications to the sensor manager.

sensor: The sensor to expect notifications from.

event: The mask of event types to expect notifications from.

Return: 0 on success, non-zero error code on failure.