Function mynewt::kernel::os::os_msys_register

pub unsafe extern "C" fn os_msys_register(arg1: *mut os_mbuf_pool) -> c_int

MSYS is a system level mbuf registry. Allows the system to share packet buffers amongst the various networking stacks that can be running simultaeneously.

Mbuf pools are created in the system initialization code, and then when a mbuf is allocated out of msys, it will try and find the best fit based upon estimated mbuf size.

os_msys_register() registers a mbuf pool with MSYS, and allows MSYS to allocate mbufs out of it.

new_pool: The pool to register with MSYS

Return: 0 on success, non-zero on failure