Module mynewt::sys::console

Display messages on Arm Semihosting Console (via OpenOCD)

Functions

buffer
Add the string to the output buffer.

console_buffer⚠
Add the string to the output buffer. C API: void console_buffer(const char *buffer, unsigned int length)

console_dump⚠
Append "length" number of bytes from "buffer" to the output buffer in hex format. C API: void console_dump(const uint8_t *buffer, unsigned int len)

console_flush⚠
Flush the output buffer to the console. C API: void console_flush(void)

console_printfloat⚠
Write a float to the output buffer, with 1 decimal place. C API: void console_printfloat(float f)

console_printhex⚠
Write a byte in hexadecimal to the output buffer. C API: void console_printhex(uint8_t v)

flush
Flush the output buffer to the console.

print
Display message msg on the Arm Semihosting console (via OpenOCD).

print_strn
Display message msg on the Arm Semihosting console (via OpenOCD).

printhex
Write a byte in hexadecimal to the output buffer. C API: void console_printhex(uint8_t v)