Module mynewt::sys::console

Display messages on Arm Semihosting Console (via OpenOCD)

Functions

buffer
Add the string to the output buffer.

console_buffer⚠
Add the string to the output buffer. C API: void console_buffer(const char *buffer, unsigned int length)

console_dump⚠
Write "length" number of bytes from "buffer" to the output buffer in hex format. C API: void console_dump(const uint8_t *buffer, unsigned int len)

console_flush⚠
Flush the output buffer to the console. C API: void console_flush(void)

console_printdouble⚠
Write a double to the output buffer, with 6 decimal places. C API: void console_printdouble(double d)

console_printfloat⚠
Write a float to the output buffer, with 2 decimal places. C API: void console_printfloat(float f)

console_printhex⚠
Write a byte in hexadecimal to the output buffer. C API: void console_printhex(uint8_t v)

console_printint⚠
Write an int i to the output buffer. C API: void console_printint(int i)

dump
Write "length" number of bytes from "buffer" to the output buffer in hex format.

flush
Flush the output buffer to the console.

get_dispatch_address
Return the Dispatch Address for the OS function whose function name hashed is hash

print
Display message msg on the Arm Semihosting console (via OpenOCD).

print_strn
Display message msg on the Arm Semihosting console (via OpenOCD).

printdouble
Write a double to the output buffer, with 6 decimal places.

printfloat
Write a float to the output buffer, with 2 decimal places.

printhex
Write a byte in hexadecimal to the output buffer. C API: void console_printhex(uint8_t v)

printint
Write an int to the output buffer. C API: void console_printint(int i)